@@ -3,8 +3,8 @@
   "version": "1.12.0",
   "description": "SendGrid API mock",
   "engines": {
-    "node": ">=16",
-    "npm": ">=8"
+    "node": ">=18",
+    "npm": ">=10"
   },
   "targets": {
     "default": {
@@ -14,12 +14,12 @@
     }
   },
   "scripts": {
-    "lint": "eslint src",
+    "lint": "eslint \"src/**/*.{js,jsx,ts,tsx}\"",
     "test": "jest",
-    "server-dev": "cross-env API_KEY=sendgrid-api-key MAIL_HISTORY_DURATION=PT24H node ./src/server/Server.js",
+    "server-dev": "cross-env API_KEY=sendgrid-api-key MAIL_HISTORY_DURATION=PT24H tsx ./src/server/Server.ts",
     "ui-dev": "parcel ./src/ui/index.html",
     "dev": "concurrently \"npm run server-dev\" \"npm run ui-dev\"",
-    "build": "parcel build ./src/ui/index.html"
+    "build": "parcel build ./src/ui/index.html && tsc --project tsconfig.json && tsc-alias -p tsconfig.json"
   },
   "author": "janjaali",
   "repository": {
@@ -41,15 +41,28 @@
     "sanitize-filename": "^1.6.3"
   },
   "devDependencies": {
-    "babel-eslint": "^10.0.2",
+    "@types/body-parser": "^1.19.5",
+    "@types/express": "^5.0.0",
+    "@types/jest": "^29.5.14",
+    "@types/json-schema": "^7.0.15",
+    "@types/sinon": "^17.0.3",
+    "@types/supertest": "^6.0.2",
     "concurrently": "^7.2.2",
     "cross-env": "^7.0.3",
-    "eslint": "^8.0.0",
-    "eslint-plugin-react": "^7.14.3",
-    "jest": "^27.0.0",
+    "eslint": "^9.14.0",
+    "eslint-plugin-react": "^7.37.2",
+    "globals": "^15.12.0",
+    "jest": "^29.7.0",
+    "json-schema": "^0.4.0",
     "parcel": "^2.6.2",
     "process": "^0.11.10",
+    "prop-types": "^15.8.1",
     "sinon": "^14.0.0",
-    "supertest": "^7.0.0"
+    "supertest": "^7.0.0",
+    "ts-jest": "^29.2.5",
+    "tsc-alias": "^1.8.10",
+    "tsx": "^4.19.2",
+    "typescript": "^5.6.3",
+    "typescript-eslint": "^8.14.0"
   }
 }
