@@ -1,8 +1,12 @@
-const bodyParser = require('body-parser');
-const crypto = require('crypto');
-const {Validator, ValidationError} = require('express-json-validator-middleware');
+import bodyParser from 'body-parser';
+import { Express, Request, Response, NextFunction } from 'express';
+import crypto from 'crypto';
+import {Validator, ValidationError} from 'express-json-validator-middleware';
+import MailHandler from '@/server/handler/MailHandler';
 
-const jsonSchema = {
+import { JSONSchema7 } from 'json-schema';
+
+const jsonSchema: Record<string, JSONSchema7> = {
   v3MailSend: {
     type: 'object',
     required: ['personalizations', 'from'],
@@ -12,7 +16,7 @@ const jsonSchema = {
       }
     },
     then: {
-      required: ['content','subject']
+      required: ['content', 'subject']
     },
     properties: {
       content: {
@@ -20,27 +24,18 @@ const jsonSchema = {
         items: {
           type: 'object',
           properties: {
-            type: {
-              type: 'string',
-            },
-            value: {
-              type: 'string',
-            },
+            type: { type: 'string' },
+            value: { type: 'string' }
           },
-        }
+        },
       },
       from: {
         type: 'object',
         required: ['email'],
         properties: {
-          email: {
-            type: 'string',
-          },
-          name: {
-            type: 'string',
-            nullable: true,
-          },
-        }
+          email: { type: 'string' },
+          name: { type: ['string', 'null'] },
+        },
       },
       personalizations: {
         type: 'array',
@@ -52,36 +47,22 @@ const jsonSchema = {
               items: {
                 type: 'object',
                 properties: {
-                  email: {
-                    type: 'string',
-                  },
+                  email: { type: 'string' }
                 },
               },
             },
-            custom_args: {
-              type: 'object',
-              nullable: true,
-            }
+            custom_args: { type: ['object', 'null'] }
           },
-        }
-      },
-      subject: {
-        type: 'string',
-      },
-      template_id: {
-        type: 'string',
-        nullable: true,
-      },
-      categories: {
-        type: 'array',
-        items: {
-          type: 'string',
         },
-        nullable: true,
+      },
+      subject: { type: 'string' },
+      template_id: { type: ['string', 'null'] },
+      categories: {
+        type: ['array', 'null'],
+        items: { type: 'string' }
       },
       custom_args: {
-        type: 'object',
-        nullable: true,
+        type: ['object', 'null']
       }
     }
   }
@@ -89,31 +70,32 @@ const jsonSchema = {
 
 /**
  * Creates a request handler for the given express app.
- * 
+ *
  * @param {*} app express app
- * @param {*} apiAuthenticationKey api key for authentication 
+ * @param {*} apiAuthenticationKey api key for authentication
  * @param {*} mailHandler  mail handler
  */
-const RequestHandler = (app, apiAuthenticationKey, mailHandler) => {
+const RequestHandler = (app: Express, apiAuthenticationKey: any, mailHandler: MailHandler) => {
 
-  const { validate } = new Validator();
+  const { validate } = new Validator({});
 
-  // Using application/json parser for parsing the request body with a slightly 
+  // Using application/json parser for parsing the request body with a slightly
   // increased limit for the request body size thus allowing larger mails.
   app.use(bodyParser.json({ limit: '5mb' }));
 
   app.post(
     '/v3/mail/send',
-    validate({ body: jsonSchema.v3MailSend }), 
+    // @ts-ignore
+    validate({ body: jsonSchema.v3MailSend }),
     (req, res) => {
 
       const reqApiKey = req.headers.authorization;
-        
+
       if (reqApiKey === `Bearer ${apiAuthenticationKey}`) {
         const messageId = crypto.randomUUID();
-          
+
         mailHandler.addMail(req.body, messageId);
-      
+
         res.status(202).header({
           'X-Message-ID': messageId,
         }).send();
@@ -129,45 +111,50 @@ const RequestHandler = (app, apiAuthenticationKey, mailHandler) => {
       }
     }
   );
-      
+
   app.get('/api/mails', (req, res) => {
-    
+
     const filterCriteria = {
-      to: req.query.to,
-      subject: req.query.subject,
-      dateTimeSince: req.query.dateTimeSince,
+      to: req.query.to?.toString(),
+      subject: req.query.subject?.toString(),
+      dateTimeSince: req.query.dateTimeSince?.toString(),
     };
-      
+
     const paginationCriteria = {
-      page: req.query.page,
-      pageSize: req.query.pageSize,
+      page: Number(req.query.page?.toString() ?? '0'),
+      pageSize: Number(req.query.pageSize?.toString() ?? '10'),
     };
-      
+
     const mails = mailHandler.getMails(filterCriteria, paginationCriteria);
-        
+
     res.send(mails);
   });
-    
+
   app.delete('/api/mails', (req, res) => {
-    
+
     const filterCriteria = {
-      to: req.query.to,
+      to: req.query.to?.toString() ?? '',
     };
-      
+
     mailHandler.clear(filterCriteria);
-      
+
     res.sendStatus(202);
   });
 
   // Error handler that returns a 400 status code if the request body does not
   // match the given json schema.
-  app.use((error, request, response, next) => {
+  app.use((
+    error: Error | ValidationError,
+    _: Request,
+    response: Response,
+    next: NextFunction
+  )  => {
     if (error instanceof ValidationError) {
       const responseBody = {
-        errors: error.validationErrors.body.map(validationError => {
+        errors: error.validationErrors.body?.map(validationError => {
           return {
-            field: validationError.params.missingProperty 
-              ? validationError.params.missingProperty 
+            field: validationError.params.missingProperty
+              ? validationError.params.missingProperty
               : validationError.propertyName,
             message: validationError.message,
             path: validationError.schemaPath,
@@ -182,4 +169,4 @@ const RequestHandler = (app, apiAuthenticationKey, mailHandler) => {
   });
 };
 
-module.exports = RequestHandler;
+export default RequestHandler;
