@@ -1,24 +1,22 @@
-const { loggerFactory } = require('./logger/log4js');
-const { setupExpressApp } = require('./ExpressApp');
-const MailHandler = require('./handler/MailHandler');
-const asHttpsServer = require('./ssl/index');
+import { loggerFactory } from './logger/log4js';
+import { setupExpressApp } from './ExpressApp';
+import MailHandler from './handler/MailHandler';
+import asHttpsServer from './ssl/index';
 
 const logger = loggerFactory('Server');
 
-const authenticationUsers = (usersString) => {
+const authenticationUsers = (usersString: string): { [x: string]: string; } => {
 
   const authenticationParams = usersString.split(';');
 
-  const users = authenticationParams.flatMap(auth => {
-
+  const users = authenticationParams.map(auth => {
     const userPasswordPair = auth.split(':');
-
-    return {
-      [userPasswordPair[0]]: userPasswordPair[1]
-    };
+    return { [userPasswordPair[0]]: userPasswordPair[1] };
   });
 
-  return Object.assign(...users);
+  return users.reduce((acc, curr) => {
+    return { ...acc, ...curr };
+  }, {});
 };
 
 const mailHandler = new MailHandler(process.env.MAIL_HISTORY_DURATION);
@@ -34,8 +32,8 @@ const rateLimitConfiguration = {
 };
 
 const app = setupExpressApp(
-  mailHandler, 
-  apiAuthentication, 
+  mailHandler,
+  apiAuthentication,
   process.env.API_KEY,
   rateLimitConfiguration,
 );
@@ -47,7 +45,7 @@ if (enableSsl) {
   logger.info('Starting send-grid mock with letsencrypt.org integration (use https)!');
 
   const sslRateLimitConfiguration = {
-    enabled: process.env.SSL_RATE_LIMIT_ENABLED === 'true', 
+    enabled: process.env.SSL_RATE_LIMIT_ENABLED === 'true',
     windowInMs: process.env.SSL_RATE_LIMIT_WINDOW_IN_MS ? process.env.SSL_RATE_LIMIT_WINDOW_IN_MS : 60000,
     maxRequests: process.env.SSL_RATE_LIMIT_MAX_REQUESTS ? process.env.SSL_RATE_LIMIT_MAX_REQUESTS : 100,
   };
